split large data model shared spec cypress file
