+++ e2e/test/scenarios/data-model/data-model.shared-3.cy.spec.ts
@@ -1,0 +1,406 @@
+import {
+  SAMPLE_DB_ID,
+  SAMPLE_DB_SCHEMA_ID,
+  WRITABLE_DB_ID,
+} from "e2e/support/cypress_data";
+import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
+
+const { H } = cy;
+const { TablePicker, TableSection, FieldSection, PreviewSection, Shared } =
+  cy.H.DataModel;
+
+const { ORDERS_ID, ORDERS } = SAMPLE_DATABASE;
+const { visitArea } = Shared;
+
+const areas: ("admin" | "data studio")[] = ["admin", "data studio"];
+type Area = (typeof areas)[number];
+
+describe.each<Area>(areas)("data model > %s", (area: Area) => {
+  const visit = visitArea(area);
+
+  beforeEach(() => {
+    H.restore();
+    H.resetSnowplow();
+    cy.signInAsAdmin();
+    H.activateToken("bleeding-edge");
+
+    cy.intercept("GET", "/api/database/*/schemas?*").as("schemas");
+    cy.intercept("GET", "/api/table/*/query_metadata*").as("metadata");
+    cy.intercept("GET", "/api/database/*/schema/*").as("schema");
+    cy.intercept("POST", "/api/dataset*").as("dataset");
+    cy.intercept("GET", "/api/field/*/values").as("fieldValues");
+    cy.intercept("PUT", "/api/field/*", cy.spy().as("updateFieldSpy")).as(
+      "updateField",
+    );
+    cy.intercept("PUT", "/api/table/*/fields/order").as("updateFieldOrder");
+    cy.intercept("POST", "/api/field/*/values").as("updateFieldValues");
+    cy.intercept("POST", "/api/field/*/dimension").as("updateFieldDimension");
+    cy.intercept("PUT", "/api/table").as("updateTables");
+    cy.intercept("PUT", "/api/table/*").as("updateTable");
+
+    if (area === "admin") {
+      cy.intercept("GET", "/api/database?*").as("databases");
+      cy.intercept("GET", "/api/field/*/values").as("fieldValues");
+      cy.intercept("PUT", "/api/table/*").as("updateTable");
+    }
+
+    if (area === "data studio") {
+      cy.intercept("GET", "/api/database").as("databases");
+    }
+  });
+
+  describe("Error handling", { tags: "@external" }, () => {
+    beforeEach(() => {
+      H.restore("postgres-writable");
+      H.resetTestTable({ type: "postgres", table: "many_data_types" });
+      cy.signInAsAdmin();
+      H.resyncDatabase({
+        dbId: WRITABLE_DB_ID,
+        tableName: "many_data_types",
+      });
+
+      const error = { statusCode: 500 };
+      cy.intercept("POST", "/api/dataset*", error);
+      cy.intercept("PUT", "/api/field/*", error);
+      cy.intercept("PUT", "/api/table/*/fields/order", error);
+      cy.intercept("POST", "/api/field/*/values", error);
+      cy.intercept("POST", "/api/field/*/dimension", error);
+      cy.intercept("PUT", "/api/table/*", error);
+      if (area === "admin") {
+        cy.intercept("POST", "/api/table/*/sync_schema", error);
+        cy.intercept("POST", "/api/table/*/rescan_values", error);
+        cy.intercept("POST", "/api/table/*/discard_values", error);
+      } else {
+        cy.intercept("POST", "/api/data-studio/table/sync-schema", error);
+        cy.intercept("POST", "/api/data-studio/table/rescan-values", error);
+        cy.intercept("POST", "/api/data-studio/table/discard-values", error);
+      }
+    });
+
+    it("shows toast errors and preview errors", () => {
+      visit({
+        databaseId: SAMPLE_DB_ID,
+        schemaId: SAMPLE_DB_SCHEMA_ID,
+        tableId: ORDERS_ID,
+        fieldId: ORDERS.QUANTITY,
+      });
+
+      cy.log("table section");
+
+      cy.log("name");
+      TableSection.getNameInput().type("a").blur();
+      verifyAndCloseToast("Failed to update table name");
+
+      cy.log("description");
+      TableSection.getDescriptionInput().type("a").blur();
+      verifyAndCloseToast("Failed to update table description");
+
+      cy.log("predefined field order");
+      TableSection.getSortButton().click();
+      TableSection.getSortOrderInput()
+        .findByLabelText("Alphabetical order")
+        .click();
+      verifyAndCloseToast("Failed to update field order");
+
+      cy.log("custom field order");
+      TableSection.getSortableField("ID").as("dragElement");
+      H.moveDnDKitElementByAlias("@dragElement", {
+        vertical: 50,
+      });
+      verifyAndCloseToast("Failed to update field order");
+      TableSection.get().button("Done").click();
+
+      cy.log("sync");
+      TableSection.getSyncOptionsButton().click();
+      H.modal().button("Sync table schema").click();
+      verifyAndCloseToast("Failed to start sync");
+
+      cy.log("scan");
+      H.modal().button("Re-scan table").click();
+      verifyAndCloseToast("Failed to start scan");
+
+      cy.log("discard field values");
+      H.modal().button("Discard cached field values").click();
+      verifyAndCloseToast("Failed to discard values");
+      cy.realPress("Escape");
+
+      cy.log("field name");
+      TableSection.getFieldNameInput("Quantity").type("a").blur();
+      verifyAndCloseToast("Failed to update name of Quantity");
+
+      cy.log("field description");
+      TableSection.getFieldDescriptionInput("Quantity").type("a").blur();
+      verifyAndCloseToast("Failed to update description of Quantity");
+
+      cy.log("field section");
+
+      cy.log("name");
+      FieldSection.getNameInput().type("a").blur();
+      verifyAndCloseToast("Failed to update name of Quantity");
+
+      cy.log("description");
+      FieldSection.getDescriptionInput().type("a").blur();
+      verifyAndCloseToast("Failed to update description of Quantity");
+
+      cy.log("coercion strategy");
+      FieldSection.getCoercionToggle().parent().scrollIntoView().click();
+      H.popover()
+        .findByText("UNIX seconds → Datetime")
+        .scrollIntoView()
+        .click();
+      verifyAndCloseToast("Failed to enable casting for Quantity");
+
+      cy.log("semantic type");
+      FieldSection.getSemanticTypeInput().click();
+      H.popover().findByText("Score").click();
+      verifyAndCloseToast("Failed to update semantic type of Quantity");
+
+      cy.log("visibility");
+      FieldSection.getVisibilityInput().click();
+      H.popover().findByText("Only in detail views").click();
+      verifyAndCloseToast("Failed to update visibility of Quantity");
+
+      cy.log("filtering");
+      FieldSection.getFilteringInput().click();
+      H.popover().findByText("Search box").click();
+      verifyAndCloseToast("Failed to update filtering of Quantity");
+
+      cy.log("display values");
+      FieldSection.getDisplayValuesInput().click();
+      H.popover().findByText("Custom mapping").click();
+      verifyAndCloseToast("Failed to update display values of Quantity");
+
+      cy.log("JSON unfolding");
+      // navigating away would cause onChange to be triggered in InputBlurChange and TextareaBlurChange
+      // components, so new undos will appear - this makes this test flaky, so we navigate with page reload instead
+      visit({ databaseId: WRITABLE_DB_ID });
+      TablePicker.getTable("Many Data Types").click();
+      TableSection.clickField("Json");
+      FieldSection.getUnfoldJsonInput().click();
+      H.popover().findByText("No").click();
+      verifyAndCloseToast("Failed to disable JSON unfolding for Json");
+
+      cy.log("formatting");
+      TablePicker.getDatabase("Sample Database").click();
+      TablePicker.getTable("Orders").click();
+      TableSection.clickField("Quantity");
+      FieldSection.getPrefixInput().type("5").blur();
+      verifyAndCloseToast("Failed to update formatting of Quantity");
+
+      cy.log("preview section");
+
+      cy.log("table preview");
+      FieldSection.getPreviewButton().click();
+      PreviewSection.get()
+        .scrollIntoView()
+        .findByText("Something went wrong")
+        .should("be.visible");
+
+      cy.log("object detail preview");
+      PreviewSection.getPreviewTypeInput().findByText("Detail").click();
+      PreviewSection.get()
+        .findByText("Something went wrong")
+        .should("be.visible");
+    });
+  });
+
+  describe("Undos", { tags: "@external" }, () => {
+    beforeEach(() => {
+      H.restore("postgres-writable");
+      H.resetTestTable({ type: "postgres", table: "many_data_types" });
+      cy.signInAsAdmin();
+      H.resyncDatabase({
+        dbId: WRITABLE_DB_ID,
+        tableName: "many_data_types",
+      });
+    });
+
+    it("allows to undo every action", () => {
+      visit({
+        databaseId: SAMPLE_DB_ID,
+        schemaId: SAMPLE_DB_SCHEMA_ID,
+        tableId: ORDERS_ID,
+        fieldId: ORDERS.QUANTITY,
+      });
+
+      cy.log("table section");
+
+      cy.log("name");
+      TableSection.getNameInput().type("a").blur();
+      verifyToastAndUndo("Table name updated");
+      TableSection.getNameInput().should("have.value", "Orders");
+
+      cy.log("description");
+      TableSection.getDescriptionInput().type("a").blur();
+      verifyToastAndUndo("Table description updated");
+      TableSection.getDescriptionInput().should(
+        "have.value",
+        "Confirmed Sample Company orders for a product, from a user.",
+      );
+
+      cy.log("predefined field order");
+      TableSection.getSortButton().click();
+      TableSection.getSortOrderInput()
+        .findByLabelText("Alphabetical order")
+        .click();
+      verifyToastAndUndo("Field order updated");
+      TableSection.getSortOrderInput()
+        .findByDisplayValue("database")
+        .should("be.checked");
+
+      cy.log("custom field order");
+      TableSection.getSortableField("ID").as("dragElement");
+      H.moveDnDKitElementByAlias("@dragElement", {
+        vertical: 50,
+      });
+      verifyToastAndUndo("Field order updated");
+      TableSection.getSortOrderInput()
+        .findByDisplayValue("database")
+        .should("be.checked");
+      TableSection.get().button("Done").click();
+
+      cy.log("field name");
+      TableSection.getFieldNameInput("Quantity").type("a").blur();
+      verifyToastAndUndo("Name of Quantity updated");
+      TableSection.getFieldNameInput("Quantity").should(
+        "have.value",
+        "Quantity",
+      );
+
+      cy.log("field description");
+      TableSection.getFieldDescriptionInput("Quantity").type("a").blur();
+      verifyToastAndUndo("Description of Quantity updated");
+      TableSection.getFieldDescriptionInput("Quantity").should(
+        "have.value",
+        "Number of products bought.",
+      );
+
+      cy.log("field section");
+
+      cy.log("name");
+      FieldSection.getNameInput().type("a").blur();
+      verifyToastAndUndo("Name of Quantity updated");
+      FieldSection.getNameInput().should("have.value", "Quantity");
+
+      cy.log("description");
+      FieldSection.getDescriptionInput().type("a").blur();
+      verifyToastAndUndo("Description of Quantity updated");
+      FieldSection.getDescriptionInput().should(
+        "have.value",
+        "Number of products bought.",
+      );
+
+      cy.log("coercion strategy");
+      FieldSection.getCoercionToggle().parent().scrollIntoView().click();
+      H.popover()
+        .findByText("UNIX seconds → Datetime")
+        .scrollIntoView()
+        .click();
+      verifyToastAndUndo("Casting enabled for Quantity");
+      FieldSection.getCoercionToggle().should("not.be.checked");
+
+      cy.log("semantic type");
+      FieldSection.getSemanticTypeInput().click();
+      H.popover().findByText("Score").click();
+      verifyToastAndUndo("Semantic type of Quantity updated");
+      FieldSection.getSemanticTypeInput().should("have.value", "Quantity");
+
+      cy.log("visibility");
+      FieldSection.getVisibilityInput().click();
+      H.popover().findByText("Only in detail views").click();
+      verifyToastAndUndo("Visibility of Quantity updated");
+      FieldSection.getVisibilityInput().should("have.value", "Everywhere");
+
+      cy.log("filtering");
+      FieldSection.getFilteringInput().click();
+      H.popover().findByText("Search box").click();
+      verifyToastAndUndo("Filtering of Quantity updated");
+      FieldSection.getFilteringInput().should(
+        "have.value",
+        "A list of all values",
+      );
+
+      cy.log("display values");
+      FieldSection.getDisplayValuesInput().click();
+      H.popover().findByText("Custom mapping").click();
+      H.modal().should("be.visible");
+      H.modal().button("Close").click();
+      verifyToastAndUndo("Display values of Quantity updated");
+      FieldSection.getDisplayValuesInput().should(
+        "have.value",
+        "Use original value",
+      );
+
+      cy.log("custom mapping");
+      FieldSection.getDisplayValuesInput().click();
+      H.popover().findByText("Custom mapping").click();
+      verifyAndCloseToast("Display values of Quantity updated");
+      H.modal().within(() => {
+        cy.findByDisplayValue("0")
+          .clear()
+          .type("XYZ", { scrollBehavior: "center" })
+          .blur();
+        cy.button("Save").click();
+      });
+      verifyToastAndUndo("Display values of Quantity updated");
+      FieldSection.get().button("Edit mapping").click();
+      H.modal().within(() => {
+        cy.findByDisplayValue("0").should("be.visible");
+        cy.findByDisplayValue("XYZ").should("not.exist");
+        cy.button("Close").click();
+      });
+
+      cy.log("foreign key");
+      TableSection.clickField("User ID");
+      FieldSection.getDisplayValuesInput().click();
+      H.popover().findByText("Use foreign key").click();
+      verifyToastAndUndo("Display values of User ID updated");
+      FieldSection.getDisplayValuesInput().should(
+        "have.value",
+        "Use original value",
+      );
+
+      cy.log("JSON unfolding");
+      TablePicker.getDatabase("Writable Postgres12").click();
+      TablePicker.getTable("Many Data Types").click();
+      TableSection.clickField("Json");
+      FieldSection.getUnfoldJsonInput().click();
+      H.popover().findByText("No").click();
+      verifyToastAndUndo("JSON unfolding disabled for Json");
+      FieldSection.getUnfoldJsonInput().should("have.value", "Yes");
+
+      cy.log("formatting");
+      TablePicker.getTable("Orders").click();
+      TableSection.clickField("Quantity");
+
+      cy.log("prefix (ChartSettingInput)");
+      FieldSection.getPrefixInput().type("5").blur();
+      verifyToastAndUndo("Formatting of Quantity updated");
+      FieldSection.getPrefixInput().should("have.value", "");
+
+      cy.log("multiply by number (ChartSettingInputNumeric)");
+      FieldSection.getMultiplyByNumberInput().type("5").blur();
+      verifyToastAndUndo("Formatting of Quantity updated");
+      FieldSection.getMultiplyByNumberInput().should("have.value", "");
+
+      cy.log("mini bar chart (ChartSettingToggle)");
+      FieldSection.getMiniBarChartToggle()
+        .parent()
+        .click({ scrollBehavior: "center" });
+      verifyToastAndUndo("Formatting of Quantity updated");
+      FieldSection.getMiniBarChartToggle().should("not.be.checked");
+    });
+  });
+});
+
+function verifyAndCloseToast(message: string) {
+  H.undoToast().should("contain.text", message);
+  H.undoToast().icon("close").click({ force: true });
+}
+
+function verifyToastAndUndo(message: string) {
+  H.undoToast().should("contain.text", message);
+  H.undoToast().button("Undo").click();
+  H.undoToast().should("contain.text", "Change undone");
+  H.undoToast().icon("close").click({ force: true });
+}
